Minor style adjustments and refactoring with SASS
