+++ src/components/global.scss
@@ -1,4 +1,4 @@
 $primary: #111D2A;
 $light: #f8f9fa;
 $dark: #343a40;
-
+$text-color-light: #f8f9fa;
