+++ src/components/home/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import MovieList from '../movie/movie-list/movie-list-container';
-import styles from './styles.css';
+import styles from './styles.scss';
 
 const Home = () => (
         <div className={styles.container}>
