+++ src/components/home/styles.scss
@@ -1,0 +1,7 @@
+h1 {
+    text-align: center;
+}
+
+.container {
+    padding: 15px 0 15px;
+}
