+++ src/components/movie/movie-card/styles.scss
@@ -3,7 +3,7 @@
 .container {
     display: inline-block;
     margin: 10px;
-    border: rgba(34,34,34, 0.5) 1px solid;
+    border: transparentize($dark, 0.3) 1px solid;
 }
 
 .title {
@@ -16,5 +16,9 @@
 }
 
 .poster {
-    width: 320px;
+    max-width: 320px;
+    img {
+      width: 100%;
+    }
 }
+
