+++ src/components/movie/movie-list/styles.scss
@@ -1,4 +1,3 @@
-
 .container {
     text-align: center;
 }
